Refactor Util namespace, move Logger to root namespace.
