+++ DSLink/Connection/KeyPair.cs
@@ -1,6 +1,4 @@
 using System;
-using System.IO;
-using System.Threading.Tasks;
 using DSLink.Util;
 using Org.BouncyCastle.Asn1.Sec;
 using Org.BouncyCastle.Crypto;
